requirements.txt
streamlit>=1.30.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.3.0
chromadb>=0.5.0
plotly>=5.18.0
pymupdf>=1.24.0
sentence-transformers>=3.0.0
langchain-huggingface>=0.1.0
python-dotenv>=1.0.0
